optimize large array storage

+++ distance.py
@@ -9,9 +9,11 @@ def compute_distance_matrix(X, Y):
     nx = X.shape[0]
     ny = Y.shape[0]
     nz = X.shape[1]
-    D = np.zeros((nx, ny), dtype=np.uint16)
+    # Pre-allocate using the smallest usable dtype, which is uint8 for our purposes for 8-bit RGB colors.
+    D = np.zeros((nx, ny), dtype=np.uint8)
     for i in prange(nx):
         for j in prange(ny):
             for k in prange(nz):
-                D[i, j] += abs(X[i, k] - Y[j, k])
+                # Use integer division to ensure that the sum does not exceed the size of np.uint8 in any entry of D
+                D[i, j] += abs(X[i, k] - Y[j, k]) // nz
     return D
